+++ NEWS.md
@@ -18,6 +18,7 @@ New preset selection:
 * Post-Install.preset - Enable Bitlocker after installation
 
 New presets/rules:
+* 2021/06/04 - DisableUWPAccessLocation/EnableUWPAccessLocation - Let UWP Apps Access Location
 * 2021/06/01 - DisableWindowsFeeds/EnableWindowsFeeds - News and Interests on Taskbar
 * 2021/05/12 - EnableBitlocker/DisableBitlocker - Crypt hard drive with bitlocker
 * 2021/04/29 - DisableWindowsHello/EnableWindowsHello - Windows Hello Authentification